Further work on debug library
- improve separation of UWP and .NET
- comment debug.writeline (isn't available in .NET)
- bumped Nuget to 0.3.0-preview002
-
